update log uploader

+++ JuggleIM/JuggleIM/internal/utility/JIntervalGenerator.h
@@ -11,7 +11,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface JIntervalGenerator : NSObject
 
-- (int)getNextInterval;
+- (float)getNextInterval;
 
 - (void)reset;
